cleanup and rearrange
